Autogenerate observables for all settings
This way we don't have to create an observable for
every setting and keep it in sync with the setting;
instead, just get/set the observable, and it is
auto-persisted.

Also, we can now get the setting's value
in multiple places without cut-and-pasting the default,
because we do a one-time registration.