Fix type of launcher generation check
This worked because for now we don't have a launcher
generation on typesafe.com so we would fall back to 0
which was an integer. But if we'd added a launcher generation,
we would have checked int-vs-string equality which would
have always failed.

Score one point in the case for a typesafe == I guess.